+++ CHANGES.rst
@@ -9,6 +9,9 @@
   beyond ASCII.
   (`#11 <https://github.com/zopefoundation/zope.app.publisher/issues/11>`_)
 
+- Support options *use_datetime* (Python >= 2.7) and *use_builtin_types*
+  (`Python >= 3.5`) in ``.xmlrpc.testing.ServerProxy``.
+
 
 4.2.0 (2019-12-05)
 ==================